Organize files (rename and move to right folders)
