+++ client/html/pages/panel/add-vacancy.html
